Ryantrois

+++ frontend/src/components/ManageRef.jsx
@@ -53,6 +53,7 @@ function ManageRef() {
     }
   }, [editPhone, editMode]);
 
+  const [showModal, setShowModal] = useState(false);
   const handleEdit = () => {
     axios
       .put(`http://localhost:5000/phone-ref/${editPhone.id}`, dataPhoneRef)
@@ -79,6 +80,7 @@ function ManageRef() {
         axios.get("http://localhost:5000/phone-ref").then((result) => {
           setPhoneRef(result.data);
         });
+        setShowModal(false);
       })
       .catch((error) => {
         console.warn(error);
@@ -155,6 +157,10 @@ function ManageRef() {
     },
   ];
 
+  const handleDeleteModal = () => {
+    setShowModal(true);
+  };
+
   return (
     <div>
       {!editMode ? (
@@ -166,18 +172,66 @@ function ManageRef() {
                 handleClick={handleClick}
                 key={phone.id}
               />
-              <button
-                className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 border border-blue-700 rounded"
-                onClick={() => handleDelete(phone)}
-              >
-                Supprimer
-              </button>
-              <button
-                className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 border border-blue-700 rounded"
-                onClick={() => handleEditMode(phone)}
-              >
-                Modifier
-              </button>
+              <div className="flex justify-around my-2">
+                <button
+                  className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 border border-blue-700 rounded"
+                  onClick={() => handleDeleteModal(phone)}
+                >
+                  Supprimer
+                </button>
+                <button
+                  className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 border border-blue-700 rounded"
+                  onClick={() => handleEditMode(phone)}
+                >
+                  Modifier
+                </button>
+              </div>
+              {showModal && (
+                <>
+                  <div className="justify-center items-center flex overflow-x-hidden mx-10 overflow-y-auto fixed inset-0 z-50 outline-none focus:outline-none">
+                    <div className="relative w-auto my-6 mx-auto max-w-3xl">
+                      <div className="border-0 rounded-lg shadow-lg relative flex flex-col w-full bg-white outline-none focus:outline-none">
+                        <div className="flex items-start justify-between p-5 border-b border-solid border-slate-200 rounded-t">
+                          <h3 className="text-xl text-center font-semibold">
+                            Voulez vous vraiment supprimer {phone.brand}{" "}
+                            {phone.model} ?
+                          </h3>
+                          <button
+                            type="button"
+                            className="p-1 ml-auto bg-transparent border-0 text-black opacity-5 float-right text-3xl leading-none font-semibold outline-none focus:outline-none"
+                            onClick={() => {
+                              setShowModal(false);
+                            }}
+                          >
+                            <span className="bg-transparent text-black opacity-5 h-6 w-6 text-2xl block outline-none focus:outline-none">
+                              ×
+                            </span>
+                          </button>
+                        </div>
+                        <div className="flex items-center justify-end p-6 border-t border-solid border-slate-200 rounded-b">
+                          <button
+                            className="text-red-500 background-transparent font-bold uppercase px-6 py-2 text-sm outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150"
+                            type="button"
+                            onClick={() => {
+                              setShowModal(false);
+                            }}
+                          >
+                            Annuler
+                          </button>
+                          <button
+                            className="bg-emerald-500 text-white active:bg-emerald-600 font-bold uppercase text-sm px-6 py-3 rounded shadow hover:shadow-lg outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150"
+                            type="button"
+                            onClick={() => handleDelete(phone)}
+                          >
+                            Supprimer
+                          </button>
+                        </div>
+                      </div>
+                    </div>
+                  </div>
+                  <div className="opacity-25 fixed inset-0 z-40 bg-black" />
+                </>
+              )}
             </div>
           ))}
         </div>
